+++ packages/peer_coordinator/lib/data.dart
@@ -1,3 +1,4 @@
+export 'src/data/clock_sync_sample.dart';
 export 'src/data/message.dart';
 export 'src/data/message_timing.dart';
 export 'src/data/network_stream.dart';